refactor(ui): extract color helpers from utils to UI layer
